Update map_sprite.cpp
[-Werror=useless-cast]

+++ src/opengl/main/map_sprite.cpp
@@ -1724,7 +1724,7 @@ std::string map_sprite::get_base_name() const
      future_of_futures.reserve(max_number_of_texture_pages);
 
      // Setup an off-screen render texture
-     iRectangle const canvas = m_map_group.maps.const_working().canvas() * static_cast<int>(m_render_framebuffer.scale());
+     iRectangle const canvas = m_map_group.maps.const_working().canvas() * m_render_framebuffer.scale();
      const auto       specification =
        glengine::FrameBufferSpecification{ .width = canvas.width(), .height = canvas.height(), .scale = settings.scale.value() };
